% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_emittimes.R
\name{write_emittimes_file}
\alias{write_emittimes_file}
\title{Write an EMITIMES file}
\usage{
write_emittimes_file(schedule, dir = ".", filename = "EMITIMES")
}
\arguments{
\item{schedule}{A list object generated by \code{set_emissions_schedule}.}

\item{dir}{The directory where the EMITIMES file should be written.
Defaults to the current working directory.}

\item{filename}{The name for the output EMITIMES file. Defaults to
"EMITIMES".}
}
\value{
Invisibly returns the full path to the written file.
}
\description{
Takes a schedule list object created by \code{set_emissions_schedule}
and writes it to a file in the HYSPLIT EMITIMES format.
}
\examples{
\dontrun{
schedule <- set_emissions_schedule(
  start_datetime = lubridate::ymd_hm("2023-08-01 00:00", tz = "UTC"),
  lat = 51.02,
  lon = 3.54,
  height = 10.0,
  rate = 1.0,
  total_emission_hours = 24
)
write_emittimes_file(schedule, dir = exec_dir, filename = "emissions_hourly.txt")
}
}
